pyst
[[This repository was converted from Subversion to git starting
2014-04-14. Conversion notes are in double square brackets as suggested
by ESR in the DVCS migration HOWTO. Comments have been edited to a
summary line plus continuation. Commit references if present are mapped
to ESRs uniform VCS-independent syntax. Note that email addresses of
committers are valid at the time of conversion to git, not necessarily
at the time of the original commit.]]
[[This commit was manufactured during conversion with a commit-date of
one second before the first real commit and includes the .gitignore
created from the default .cvsignore]]